requirements.txt
langchain-openai
python-dotenv
huggingface_hub
streamlit